requirements.txt
dotenv
python-dotenv
logging
azure-ai-inference
azure-core
aiogram
colorlog
